+++ packages/plugins/Trader/src/SiteAdaptor/trader/TraderStateBar.tsx
@@ -20,8 +20,8 @@ import { useChainContext, useEnvironmentContext, useNetworkContext, useWeb3Other
 import { isLessThan, leftShift, multipliedBy, rightShift } from '@masknet/web3-shared-base'
 import { ChainId, formatWeiToEther, SchemaType, ZERO_ADDRESS } from '@masknet/web3-shared-evm'
 import { MINIMUM_AMOUNT, MIN_GAS_LIMIT } from '../../constants/trader.js'
-import { isNativeTokenWrapper } from '../../helpers/trader.js'
-import { resolveTradeProviderName } from '../../pipes.js'
+import { isNativeTokenWrapper } from '../../helpers/isNativeTokenWrapper.js'
+import { resolveTradeProviderName } from '../../helpers/resolveTradeProviderName.js'
 import { AllProviderTradeContext } from '../../trader/useAllProviderTradeContext.js'
 import { useTradeApproveComputed } from '../../trader/useTradeApproveComputed.js'
 import type { TradeComputed, TraderAPI } from '@masknet/web3-providers/types'
