+++ packages/plugins/Trader/src/SiteAdaptor/trending/PriceChartDaysControl.tsx
@@ -1,7 +1,7 @@
 import { Link, Stack, Typography } from '@mui/material'
 import { makeStyles } from '@masknet/theme'
 import { Days } from '@masknet/shared-base'
-import { resolveDaysName } from '../../pipes.js'
+import { resolveDaysName } from '../../helpers/resolveTradeProviderName.js'
 
 const useStyles = makeStyles()((theme) => ({
     root: {
