--- packages/plugins/Trader/src/helpers/index.ts
@@ -1,3 +1,0 @@
-export * from './pair.js'
-export * from './uniswap.js'
-export * from './trader.js'
