+++ packages/plugins/Trader/src/helpers/isTradeBetter.ts
