+++ packages/plugins/Trader/src/helpers/resolveTradeProviderName.ts
@@ -27,12 +27,3 @@ export const resolveTradeProviderName = createLookupTableResolver<TradeProvider,
         throw new Error(`Unknown provider type: ${tradeProvider}`)
     },
 )
-
-export function resolveDaysName(days: number) {
-    if (days === 0) return 'MAX'
-    if (days >= 365) return `${Math.floor(days / 365)}Y`
-    if (days >= 30) return `${Math.floor(days / 30)}M`
-    if (days >= 7) return `${Math.floor(days / 7)}W`
-    if (days === 1) return '24H'
-    return `${days}d`
-}
