+++ packages/plugins/Trader/src/trader/useBalancer.ts
@@ -7,7 +7,7 @@ import { Balancer } from '@masknet/web3-providers'
 import type { ChainId, NetworkType } from '@masknet/web3-shared-evm'
 import type { TraderAPI } from '@masknet/web3-providers/types'
 import { useSlippageTolerance } from './useSlippageTolerance.js'
-import { getEVMAvailableTraderProviders } from '../utils.js'
+import { getEVMAvailableTraderProviders } from '../getEVMAvailableTraderProviders.js'
 
 export function useBalancer(
     inputAmount_: string,
